refactor(ui): split signup.dart file into three separate files
