+++ lib/main.dart
@@ -2,6 +2,8 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:indicai/pages/login/login.dart';
 import 'package:indicai/pages/signup/signup.dart';
+import 'package:indicai/pages/signup/signup-vendor.dart';
+
 import 'pages/splashScreen.dart'; 
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
